more moving around of components
