+++ src/components/CiimView/HeaderComponents/ActionButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './ActionButton';
