+++ src/components/CiimView/HeaderComponents/Divider/index.js
@@ -1,0 +1,1 @@
+export { default } from './Divider';
