+++ src/components/CiimView/HeaderComponents/ToggleFilterPaneButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './ToggleFilterPaneButton';
