+++ src/components/CiimView/HeaderComponents/index.js
@@ -1,7 +1,1 @@
-
-// eslint-disable-next-line import/prefer-default-export
-export { default as SearchBar } from './SearchBar';
-export { default as Divider } from './Divider';
-export { default as ActionButton } from './ActionButton';
-export { default as ToggleFilterPaneButton } from './ToggleFilterPaneButton';
-export { default as Header } from './Header';
+export { default } from './Header';
